Update parameterization algorithm (#156)
This commit implements an updated parameterization algorithm for the AnalysisGraph that, upon not finding exact matches for particular set of parameterization constraints, goes through a user-configurable list of 'axes' (with some sensible defaults - `["year", "month"]`) until it finds matches, and then aggregates them according to a user-specifiable aggregation function (defaults to numpy.mean). It also deals with the case where there are multiple indicators with the same name, but different units.

This should ensure that none of the indicator mean values are null.