Parameterize unit selection (#280)
The parameterize() function in AnalysisGraph.py can now handle unit selection. The unit argument takes in a dictionary object that has the indicator variables as keys and the corresponding units as values. The get_indicator_value() function in assembly.py will throw a warning and uses its default setting for units if incorrect units are passed to its unit argument.

The backcasting demo notebook now has an example of assigning units to a specific indicator variable.

* Solved Merge conflicts

* Solved Merge Conflict

* parameterizes now allows users to select units

* Solving issue where the wrong units are requested

* Removed unneccessary function call

* Can now see list of indicator variables in CAG

* get_indicator_value can now handle getting passed wrong units

* Reformatting

* CHanges reformatted

* Changes Reformatted

* Includes example of unit selection

* Ignore scripts/data

* Made mistake, fixed ignoring scripts/data