Updated AnalysisGraph::sample_from_posterior() to not return a value and rest of the code to just use this->A_original after calling it.
I was also trying to update AnalysisGraph::calculate_log_likelihood to not return a value and set this->log_likelihood within the method. However after doing it, the prediction results graph became different. Insteda of the cone shaped uncertaninty, it became uniform. Still could not figure out the reason for this behavior. I have left the modification that led to this behavior commented within AnalysisGraph::calculate_log_likelihood.