Fixed the plotting functions
1. Incorporated a small plotting bug fix I did in the hairdryer model
branch.

2. Fixed a warning caused by newer Matplotlib (3.3.0+) gave possibly due
to lack of backward compatibility. After fixing this warning the major
problem with x-axis not being properly rendered in the plots went away.

3. Pandas concat function deprecated the join_axes keyword argument in
version 0.25.0. The new version 1.1.1 does not have that keyword
argument which caused an error. This too is not backward compatible.
Removed join_axes keyword argument.

4. Although x-axis in plots represents years and months, we were just
passing in a sequence of text labels. This prevented plotting library to
treat x-axis as a date axis. This lead to incorrect graphs when there
were missing values for some dates. Changed x-axis to be a date-time
axis.

5. There were another warning, which I could not figure out how to fix.
For now it did not seem to effect the graphs generated.

While wrestling with this code, I got the feeling that there could be a
lot of repeated code. It is worth a try to eliminate repetitions if any.