Reduced the code repetition in evaluation.py::pred_plot() function.
This function had a lot of repeated code. This made maintaining this
code very difficult. Combined repeated code as much as possible. Before
condensing code the function consumed 545 lines (including comments) in
the file. After refactoring it got reduced to 203 lines (including
comments).

While refactoring tried to improve the features of the function.
Now when we plot Comparison plots and the true data is plotted, if there
are multiple points coincide on top of each other, the size of the
scatter plot markers are increased proportional to the number of
coinciding points. Also, the color of the marker changes according to
the number of coinciding points.