Cpp cleanup 2 (#469)
* 1. Updated the make target clean so that it does not complain about
files not being present when trying to remove them.

2. Updated the CMakeLists.txt so that the DelphiPython library gets
automatically copied to delphi/cpp folder as part of the build process.

* 1. Changed some loop variable names and variables names to better
reflect what they stand for.

2. Updated some comments and corrected a spelling mistake.

3. Added some TODO notes.

* Commented out some code that is not beeing used at the moment

* 1. Commented out some more code that is not being used at the moment.

2. Updated some comments: deleted obsolete comments, corrected spelling
mistakes, added more descriptive comments.

* Fixed the plotting functions

1. Incorporated a small plotting bug fix I did in the hairdryer model
branch.

2. Fixed a warning caused by newer Matplotlib (3.3.0+) gave possibly due
to lack of backward compatibility. After fixing this warning the major
problem with x-axis not being properly rendered in the plots went away.

3. Pandas concat function deprecated the join_axes keyword argument in
version 0.25.0. The new version 1.1.1 does not have that keyword
argument which caused an error. This too is not backward compatible.
Removed join_axes keyword argument.

4. Although x-axis in plots represents years and months, we were just
passing in a sequence of text labels. This prevented plotting library to
treat x-axis as a date axis. This lead to incorrect graphs when there
were missing values for some dates. Changed x-axis to be a date-time
axis.

5. There were another warning, which I could not figure out how to fix.
For now it did not seem to effect the graphs generated.

While wrestling with this code, I got the feeling that there could be a
lot of repeated code. It is worth a try to eliminate repetitions if any.

* Reduced the code repetition in evaluation.py::pred_plot() function.

This function had a lot of repeated code. This made maintaining this
code very difficult. Combined repeated code as much as possible. Before
condensing code the function consumed 545 lines (including comments) in
the file. After refactoring it got reduced to 203 lines (including
comments).

While refactoring tried to improve the features of the function.
Now when we plot Comparison plots and the true data is plotted, if there
are multiple points coincide on top of each other, the size of the
scatter plot markers are increased proportional to the number of
coinciding points. Also, the color of the marker changes according to
the number of coinciding points.

* Updated evalustions.py::pred_plot() method a bit more

1. Removed some repeated condition checks by moving some code into
another place where the same condition is being checked.

2. I noticed that for the Error plot, the confidence interval is not
shown. I investigated about. After investigating a while I realized that
when there are NaN values in the dataframe, the fill between plot does
not get generated. Updated the code to filter out NaN rows before
passing the values to the plotting function.

Co-authored-by: Manujinda Wathugala <manujinda@email.arizona.edu>