Changed some of the variables to refer to theta.
While doing that I found that the variable sampled_thetas in function
AnalysisGraph::get_edge_weights_for_causemos_viz() (file:
causemos_intergration.cpp) is defined but never used. Since I did not
write this code, I did not remove it. Was it intended to be used someway
and by mistake not used?

It seems we can keep the rest of the variables to mention beta.

The only thing that should be changed is the method name
construct_beta_pdfs() to construct_theta_pdfs()