Brought back in the discretized version of the solution for a system of differential equations as an option.
1. Now train model has an additional parameter, use_continuous. The
default is true where delphi uses continuous version of the solution to
the system of differential equation, which uses matrix exponential. When
false the discretized version is used as the solution.

2. The discretized solution was brought in without modifying the code for
the continuous version. So there is a good deal of repeated code. In the
next step, I will try to combine the repeated code.

3. Updated some comments.

4. Removed some print statements.
differential euation