Delphi fix math (#470)
* Some code cleaning

1. Deleted some of the commented code blocks.

2. Corrected some spelling mistakes in comments.

* Corrected the theta to beta conversion problem.

We have been using thetas in all the equations. However, before
computing the transition matrix, we should have converted thetas to
betas (beta = tan(theta)). Performed this update.

Lots of variable names and comments mentions beta when it is really
theta. To  make the code semantically read correctly, we should update
these names as much as possible.

* Changed the Edge member variable name to theta from its original name
beta.

Changed rest of the code accordingly. There are other variables and
comments that mention beta, which needs to be gradually changed to
theta. There are some places where it could be fine to leave it as beta.

* Changed some of the variables to refer to theta.

While doing that I found that the variable sampled_thetas in function
AnalysisGraph::get_edge_weights_for_causemos_viz() (file:
causemos_intergration.cpp) is defined but never used. Since I did not
write this code, I did not remove it. Was it intended to be used someway
and by mistake not used?

It seems we can keep the rest of the variables to mention beta.

The only thing that should be changed is the method name
construct_beta_pdfs() to construct_theta_pdfs()

* Changed the name of the method AnalysisGraph::construct_beta_pdfs() to
AnalysisGraph::construct_theta_pdfs()

* Changed the fine name of construct_beta_pdfs.cpp to
construct_theta_pdfs.cpp

* Changed the transition matrix to the continuous form.

Although we changed our LDS model calculations to matrix exponential, we
had been using the discretized version of the transition matrix, which
is incorrect.

Should ask Adarsh to review this change and check everything is in
order.

* 1. Fixed a spelling mistake

2. Updated a comment

* Fixed some spelling mistakes

* Brought back in the discretized version of the solution for a system of
differential equations as an option.

1. Now train model has an additional parameter, use_continuous. The
default is true where delphi uses continuous version of the solution to
the system of differential equation, which uses matrix exponential. When
false the discretized version is used as the solution.

2. The discretized solution was brought in without modifying the code for
the continuous version. So there is a good deal of repeated code. In the
next step, I will try to combine the repeated code.

3. Updated some comments.

4. Removed some print statements.
differential euation

* 1. Reduced some of the repeated code.

To reduce code repetition, introduced a new private method
AnalysisGraph::set_log_likelihood_helper().

2. Removed some comments.

* I think I fixed our matrix exponential calculation

We have been using A.pow(ts) as the matrix exponential. After reading
the Eigen manuals
(https://eigen.tuxfamily.org/dox/unsupported/group__MatrixFunctions__Module.html)
(unfortunately they are not clear either. All the formulas and equations
are missing), I felt that using pow() is incorrect. I felt the correct
function to use is exp().

We have to calculate e^At. First we calculate At = A * t. Then we
exponentiate the result, At.exp().

I am not exactly sure whether this is correct. However, after this
change prediction plots generated using continuous version and
discretized version are comparable. I have to get Adarsh's opinion about
this.

Co-authored-by: Manujinda Wathugala <manujinda@email.arizona.edu>