Updating the create_model portion of CauseMos integration completed
1. Now Delphi extracts observation sequence for each indicator specified
in the json request and populates the observation data structure.

2. At the moment we assume that there is one indicator per concept and there is
a single observation per indicator per time point. However, most of the
new code is capable of handling multiple indicators per concept and
multiple observations per indicator per time point. If we decided to
expand the API likewise, there are some legacy code that needs to be
updated accordingly.

3. The json specification does not specify the observation frequency,
observation order or whether observations for different indicators are
time aligned. The code handles most of these ambiguities. Also, we
assume that the observation frequency is monthly. There are some code in
place to infer a frequency less than monthly that aligns observation
sequences from the provided data. The advantage of such scheme is less
missing data points. However, to reach that level, we would have to
update some of the older code in other functions. One such method is
AnalysisGraph::calculate_num_timesteps(), which assumes a monthly
frequency when calculating the number of time steps. We also have to
update the plotting functions. There could be other functions I do not
foresee that needs updating.

4. The function that does all the work getting observations from json
and populating internal observations data structure is, for now, a
single long function. That needs to be refactored into smaller coherent
functions.

5. Now we have two separate methods that sets the internal member
observed_state_sequence, which is not good programming practice. This
could make code maintenance harder. It would be better if we could
combine them together.

6. Also we have two places that perform parametarization, where we set
indicator means and standard deviations. It would be nice to merge them
together.

7. The AnalysisGraph::from_causemos_json_dict() does not set the
indicator standard deviation. This could be an omission, which I guess
needs to be included.