Refactored AnalysisGraph::set_observed_state_sequence_from_json_dict() method into two other methods:
1. AnalysisGraph::extract_concept_indicator_mapping_and_observations_from_json()
This extracts observation sequences from json input and puts them in a
data structure while aligning time points.

2. AnalysisGraph::assess_observation_frequency()
Looking at the number of months in between consecutive observation
series of each indicator, this method assesses the effective frequency
for all the observations.
At the moment, we assume a monthly frequency and the assessment in this
method is not utilized. I coded this now, when I had the idea so that
the code is there so that later if we decide to move into a indicator
observation frequency adaptive modeling framework, we have this
handy at hand.

3. Some of the methods should be made private.

4. Need to add comments.