Fixed a bug in querying delphi.db
When querying delphi.db for observations for a particular month in the
AnalysisGraph::get_data_value() method defined in data.cpp, the
query string used checked in the 'Year' column for the month we are
searching for data. Thus in many cases, Delphi was unable to find the
data, which is already there in the database. Fixed that error.

Also, the guiding messages printed in this method was broken. Fixed
them.

Moreover in AnalysisGraph::map_concepts_to_indicators() method defined
in indicator_manipulation.cpp, added a new line to a message to make
displayed messages more readable.