Code cleaning and re-organizing
I was asked to document the Delphi Software Architecture. To do that I
went through all the code at a high level. While doing that I updated
comments, moved few methods to better locations and removed some
commented code.

Although this branch does not reflect what I had done under this commit,
since I am the only person working on the code, this branch has the
latest code and this branch is not ready to merge with the master, I did
the modification in this branch itself.

While my review, I noticed some methods that should be private have been
declared as public. This makes the AnalysisGraph interface muddier. We
need to clean up the interfaces later some time.