Delphi update rest api (#471)
* Populating observation sequence from CauseMos json file (create_model) in progress. Not yet complete

* Updating the create_model portion of CauseMos integration completed

1. Now Delphi extracts observation sequence for each indicator specified
in the json request and populates the observation data structure.

2. At the moment we assume that there is one indicator per concept and there is
a single observation per indicator per time point. However, most of the
new code is capable of handling multiple indicators per concept and
multiple observations per indicator per time point. If we decided to
expand the API likewise, there are some legacy code that needs to be
updated accordingly.

3. The json specification does not specify the observation frequency,
observation order or whether observations for different indicators are
time aligned. The code handles most of these ambiguities. Also, we
assume that the observation frequency is monthly. There are some code in
place to infer a frequency less than monthly that aligns observation
sequences from the provided data. The advantage of such scheme is less
missing data points. However, to reach that level, we would have to
update some of the older code in other functions. One such method is
AnalysisGraph::calculate_num_timesteps(), which assumes a monthly
frequency when calculating the number of time steps. We also have to
update the plotting functions. There could be other functions I do not
foresee that needs updating.

4. The function that does all the work getting observations from json
and populating internal observations data structure is, for now, a
single long function. That needs to be refactored into smaller coherent
functions.

5. Now we have two separate methods that sets the internal member
observed_state_sequence, which is not good programming practice. This
could make code maintenance harder. It would be better if we could
combine them together.

6. Also we have two places that perform parametarization, where we set
indicator means and standard deviations. It would be nice to merge them
together.

7. The AnalysisGraph::from_causemos_json_dict() does not set the
indicator standard deviation. This could be an omission, which I guess
needs to be included.

* Refactored AnalysisGraph::set_observed_state_sequence_from_json_dict()
method into two other methods:

1. AnalysisGraph::extract_concept_indicator_mapping_and_observations_from_json()
This extracts observation sequences from json input and puts them in a
data structure while aligning time points.

2. AnalysisGraph::assess_observation_frequency()
Looking at the number of months in between consecutive observation
series of each indicator, this method assesses the effective frequency
for all the observations.
At the moment, we assume a monthly frequency and the assessment in this
method is not utilized. I coded this now, when I had the idea so that
the code is there so that later if we decide to move into a indicator
observation frequency adaptive modeling framework, we have this
handy at hand.

3. Some of the methods should be made private.

4. Need to add comments.

* Added a lot of descriptive comments about the new methods.

* Changed the name of the method:
AnalysisGraph::assess_observation_frequency() to
AnalysisGraph::infer_least_common_observation_frequency()
to reflect what it does better.

* Changed method from static to normal:

AnalysisGraph::from_causemos_json_dict() was unnecessarily declared as
static. Changed that to a normal method in AnalysisGraph.

* Added a new method AnalysisGraph::generate_create_model_response() to
generate the response for the create model request by HMI. This replaces
the method AnalysisGraph::get_edge_weights_for_causemos_viz(), which we
can remove after properly testing this method.

For the time being, this method always reports success as the status. We
need to update this to convey the create model errors properly to the
HMI.

* Fixed a bug in querying delphi.db

When querying delphi.db for observations for a particular month in the
AnalysisGraph::get_data_value() method defined in data.cpp, the
query string used checked in the 'Year' column for the month we are
searching for data. Thus in many cases, Delphi was unable to find the
data, which is already there in the database. Fixed that error.

Also, the guiding messages printed in this method was broken. Fixed
them.

Moreover in AnalysisGraph::map_concepts_to_indicators() method defined
in indicator_manipulation.cpp, added a new line to a message to make
displayed messages more readable.

* Added a flag to distinguish when Delphi is called through the CauseMos
interface.

Starting to work on unifying parameterization code from CauseMos call
and local delphi call.

* Code cleaning and re-organizing

I was asked to document the Delphi Software Architecture. To do that I
went through all the code at a high level. While doing that I updated
comments, moved few methods to better locations and removed some
commented code.

Although this branch does not reflect what I had done under this commit,
since I am the only person working on the code, this branch has the
latest code and this branch is not ready to merge with the master, I did
the modification in this branch itself.

While my review, I noticed some methods that should be private have been
declared as public. This makes the AnalysisGraph interface muddier. We
need to clean up the interfaces later some time.

* Moved two methods to different source files for better organization

1. Moved AnalysisGraph::sample_transition_matrix_collection_from_prior()
from sampling.cpp to causemos_integration.cpp, which is the only place
it is being called.

2. Moved AnalysisGraph::format_projection_result(), from inference.cpp
to causemos_integraton.cpp, which is the only place it is being called.

* Changed the name of a method to better reflect what it is doing:

Changed AnalysisGraph::sample_predicted_latent_state_sequences() to
AnalysisGraph::generate_predicted_latent_state_sequences_from_sampled_parameters().

* Renamed files defining classes to consistency

Capitalized the fine names of the source files defining classes.

kde -> KDE
tran_mat_cell -> Tran_Mat_Cell
random_variable -> Random_Variable

* Updated include statements according to changed file names.

* Added a comment

Co-authored-by: Manujinda Wathugala <manujinda@email.arizona.edu>