Re-implemented the AnalysisGraph::parameter_initialization() method
1. Earlier it was coded to be called only when delphi is run locally.
Then it for each indicator read the set of first observations and
computed the mean of them. Now it uses the observation_state_sequence
and set the indicator mean according to the aggregation_method. Now it
can be called both on local call and CauseMos call.

2. There are few issues to be solved.
What to do when the first (last) observation is missing and the
aggregation_method is first (last)?

3. I set the indicator standard deviation to the maximum standard
deviation of the whole observation sequence. Note that since for each
time step, there could be multiple observations, we can calculate the
standard deviation at each time step. I am talking of the maximum of
this standard deviation time series. Is this a good idea? Also when we
have a single observation per time step observation sequence, the
maximum standard deviation becomes NaN. In such cases by default I am
setting the standard deviation to 1. Is that a good idea?

4. Implemented a method in utils.hpp to compute standard deviation.

5. While implementing that I noticed that all the other functions (e.g.
mean, median, etc) takes in vectors by copy, which is a waste. I
redefined those functions to get a const reference of the vectors.