Renamed a method
get_data_value() -> get_observations_for()

Also I realized that in get_observations_for() (earlier get_data_value()),
when we do not have monthly observations, we assume that the yearly
observation is an aggregate of monthly values and divide it by 12. This
is incorrect math if the yearly observation is not a sum of monthly
values. We need a more intelligent way to handle this situation.