Implemented applying constraints (interventions or clamping latent states)
Modified AnalysisGraph::generate_latent_state_sequences() to consider
any latent state constraints present. To do that changed the matrix
exponential (continuous) version as follows:

Before this change we were calculating:
s_t = e^At * s_0

Now we are calculating
s_t = e^A * s_{t-1}

This changes allows us to merge the two cases matrix exponential and
discretized versions nicely into one. We just have to perform e^A for
the matrix exponential version before using it int the equation. In this
commit this reduction in code is not yet done. It is the next thing that
will be done and next commit will be about that.