Added code to parse the create-experiment::projection request json
The code:

1. Checks for missing fields in the json. At the moment it does not
report any errors and the response API does not have any provisions for
this.

2. Extracts start year, start month, end year, end month and number of
prediction time steps. Since Delphi is somewhat hard coded to operate in
monthly frequency due to the earlier CauseMos API, for now things are
converted to year-month format from Posix time stamps. We assume that
the prediction is also in monthly frequency. See code for detailed
comments how we are handling this at the moment.

3. Extracts constraints provided and populates the data structure for
one-off constraints. Constraints are process in conjunction with the
first indicator attached to each concept.