Fixed a bug
I was calling some methods out of order. So, when I call
AnalysisGraph::initialize_parameters(), the member variable n_timesteps
is not set to the number of observation time steps leading the code to
seek the maximum of an empty vector crashing the program.

I have corrected the order of calls.