Refactored create projection experiment related code from runExperiment() into a separate function runProjectionExperiment() further reducing the clutter.
Added placeholders for other experiment types.

While testing the code, I experienced a bit off a weird behavior which
is undesirable but I could not figure out the reason.
I saw some weird behavior when we request results for an invalid
experiment ID just after running an experiment. The trained model seemed
to be not saved to the database. The model got re-trained from scratch
on a subsequent experiment after the initial experiment and the invalid
experiment result request. When I added a sleep between the initial
create experiment and the invalid result request this re-training did not
occur.