I think I figured out the reason for the model being trained for a second time.
Since the first request for results fails, it returns quite fast.
Therefore, the subsequent request for results is received by the Flask
server before the previous model training gets completed and written to
the database. So, the second call also retrieves the untrained model
from the database and trains it.