Fixed generating the predicted observation sequence.
When generating the predicted observation sequence we were adding
Gaussian noise, which was incorrect. Changed it so that:

predicted observed value = predicted latent value * scaling factor

After fixing this a bug in a library we are using for plotting got
revealed. When all the values are same, scipy stats.bayes_mvs returns NaN
as the mean, which is incorrect. Due to this in such cases plots do not
display anything.