Fixed a bug in constructing theta pdfs
When constructing theta pdfs we query the table gradableAdjectiveData
in delphi.db and construct the adjective response map. The map maps each
adjective to a vector of responses. If an adjective is not in the map we
should create a list with a single element mapped onto that adjective.
If an adjective is already in the map we just have to push back the new
response to the already mapped vector.

Instead, the if condition was set to do the total opposite and hence
each adjective ended up being mapped to a vector with a single element,
which is the last response got from the query for each adjective.