Delphi causemos bug fix 1 (#480)
* Allowed creating a model without any data

When called through CauseMos create-model route, allowed creating a model without providing
any training data. Blocked the ability to train such a model.

* Fixed a clamping related bug

Clamping was happening one time step after the requested time step.
Fixed it.

* Fixed generating the predicted observation sequence.

When generating the predicted observation sequence we were adding
Gaussian noise, which was incorrect. Changed it so that:

predicted observed value = predicted latent value * scaling factor

After fixing this a bug in a library we are using for plotting got
revealed. When all the values are same, scipy stats.bayes_mvs returns NaN
as the mean, which is incorrect. Due to this in such cases plots do not
display anything.

* Fixed a bug in constructing theta pdfs

When constructing theta pdfs we query the table gradableAdjectiveData
in delphi.db and construct the adjective response map. The map maps each
adjective to a vector of responses. If an adjective is not in the map we
should create a list with a single element mapped onto that adjective.
If an adjective is already in the map we just have to push back the new
response to the already mapped vector.

Instead, the if condition was set to do the total opposite and hence
each adjective ended up being mapped to a vector with a single element,
which is the last response got from the query for each adjective.

Co-authored-by: Manujinda Wathugala <manujinda@email.arizona.edu>