1. Optimized the code for speed. Previously when training we computed a matrix exponential for each observation point to evolve the system to that timestep. Now, we compute the gaps between each pair of adjacent observation points and pre-compute all the required matrix exponentials to advance the system for each observation point to the next preventing computing it repeatedly. This improved the speed roughly 3 fold.
2. Removed commented out and debug code

3. Fixed a bug
Delphi was returning nan's as predictions when predictions were done on
an already trained model. Found out that when loading the trained model
from the json dump of the model, we are not setting the indicator means
and since we do not train it again, they remained set to zero. This
caused a division by zero when processing clamping constraints.