Cleanup
1. Removed some unused methods and commented out code.

2. Added and updated some comments

3. Now the delphi local run and CauseMos run are diverged. The CauseMos
run can only use the continuous (matrix exponential) version to train
the model. For the moment the local run still uses year month format and
it can only use to discrete transition matrix format to train the model.

TODO:
The local run needs to be updated to match the CauseMos run which uses
the matrix exponential. Afterwards we will not be able to use the
discretized transition matrix to train the model.