Delphi year month to epoch (#485)
* Made causemos integration use epoch instead of year and month to index observations. Need to update the training and experiment portions to match with this.

* Undo changes in these files. Only for debugging.

* Calculative gaps 2nd approach

* Added two tests files with dummy ephoch values to make testing and debugging easy

* Calculative gaps 2nd approach

* Earlier I have committed an empty file. This is the one with the data

* Refactored the train_model method

* delta_t changes

* Year-Month to epoch index conversion for causemos integration completed

* Saved this->n_timesteps, the number of training times steps, to json output and load it back from json

* 1. Deleted commented code
2. Updated and added comments
3. Fixed two bugs in calculations

Although year-month to epoch transfer of the code is mostly complete,
when observations are greatly misaligned, aligning them creates a lot of
missing data points in between. For example a test file with about 360
observations ends up being about 8500 observations due to 23, 24, and 25
missing data points between observations after time aligning the data.

* Re-implemented the observation alignment algorithm.

1. The previous implementation ended up being slow.
2. Now
   a) Sort all the epochs where any observations are available
   b) Compute the number of epochs between each adjacent pair of
   observations.
   c) Find the most frequent gap between adjacent observations. If more
   than on gap has the same highest frequency, get the smallest of
   those.
   d) Choose that are the modeling period (or observation sampling
   period, number of epochs per single modeling delta t).
   e) Convert observation epochs to modeling timestamps based on the
   selected modeling period.

* 1. Optimized the code for speed.
Previously when training we computed a matrix exponential for each
observation point to evolve the system to that timestep. Now, we compute
the gaps between each pair of adjacent observation points and pre-compute
all the required matrix exponentials to advance the system for each
observation point to the next preventing computing it repeatedly. This
improved the speed roughly 3 fold.

2. Removed commented out and debug code

3. Fixed a bug
Delphi was returning nan's as predictions when predictions were done on
an already trained model. Found out that when loading the trained model
from the json dump of the model, we are not setting the indicator means
and since we do not train it again, they remained set to zero. This
caused a division by zero when processing clamping constraints.

* Some code and test files to help debugging

* Cleanup

1. Removed some unused methods and commented out code.

2. Added and updated some comments

3. Now the delphi local run and CauseMos run are diverged. The CauseMos
run can only use the continuous (matrix exponential) version to train
the model. For the moment the local run still uses year month format and
it can only use to discrete transition matrix format to train the model.

TODO:
The local run needs to be updated to match the CauseMos run which uses
the matrix exponential. Afterwards we will not be able to use the
discretized transition matrix to train the model.

* Terminology change for plotting

Changed Confidence Interval to Credibility Interval

Co-authored-by: Aishwarya Radhakrishnan <aishwarya.rk347@gmail.com>
Co-authored-by: Manujinda Wathugala <manujinda@email.arizona.edu>