Updated setting initial derivative and beta
1. Added the options to sample the initial derivatives and betas from
the prior. Earlier initial derivatives were set to 0 and initial betas
were set to values like mean or median of the prior or 0.

Experiments with the uncharted generated test data showed that for that
one data set setting initial derivatives to 0 and initial betas to median
of the prior, made the sampler converge better.