Delphi no data model (#486)
* ”Train” a model without any data.

1. Model is “trained” by sampling from the prior distributions.

2. Updated model/prediction results plotting.

3. Made N(0, 0.1) the prior for derivatives. We were using a uniform
prior.

4. Made the default indicator mean (should be corrected as scaling
factor) 1. Previously it was set to 0.

5. Added some test data and updated debugger.py. We can incorporate some
of them with the automated testing.

* Updated setting initial derivative and beta

1. Added the options to sample the initial derivatives and betas from
the prior. Earlier initial derivatives were set to 0 and initial betas
were set to values like mean or median of the prior or 0.

Experiments with the uncharted generated test data showed that for that
one data set setting initial derivatives to 0 and initial betas to median
of the prior, made the sampler converge better.

* Exposed beta and derivative initial value setup in the CauseMos
integration API to make experiment with different initializing schemes
easier.

Co-authored-by: Manujinda Wathugala <manujinda@email.arizona.edu>