Fixed a bug with python fork
1. How a new process should be created (spawn, fork etc.) should only be
specified once per script. The way I had included this call makes that
line executed multiple times. Fixed it.

2. Figured out how to pass the Flask app context to a user created
process and updated the code to do so. Now we do not have to rely on C++
to update the database. We can use SQLAlchemy and do it within the user
created process.

3. Made create experiment utilize a user created process to make that
process asynchronous and to quickly send the create model response to
the user.

4. TO TEST THE CREATE MODEL ASYNCHRONOUS BEHAVIOR, ARTIFICIALLY SLOWED
DOWN THE AnalysisGraph::generate_latent_state_sequences() method in
prediction.cpp BY INTRODUCIG A SLEEP(). *** REMOVE THAT BEFORE MERGING
TO MASTER ***.