MAP Sample
1. Added code to identify and retain the MAP sample even if we encounter
during the burning period.

2. If we encounter the MAP sample within the burning period it is always
put at res - 1 position of the retained samples and we reduce the number
of samples retained during the sample resolution period by 1.

3. Finally delphi remembers the index of the MAP sample withing the
retained samples.

Note: There are some (or more) repeated code in this push and previous
ones that needs to be cleaned up when there is more time.