Fixed a bug
The seasonality was not visible when Delphi was invoked through the
CauseMos interface. When a model is created, each concept’d period for
seasonality is inferred from data. However, it was not saved into the
serialized model.

When creating an experiment, an AnalysisGraph object is created by
deserializing the saved model. Since concept periods were not saved all
the concepts ended up having the default period of 1.