Delphi modeling 1 dev (#493)
* Accumulating log likelihoods

1. Remember the sequence of log likelihoods throughout the sampling
process including burning.

2. Complete state now includes the log likelihood sequence.

3. Plotting function now plots the log likelihoods.

* MAP Sample

1. Added code to identify and retain the MAP sample even if we encounter
during the burning period.

2. If we encounter the MAP sample within the burning period it is always
put at res - 1 position of the retained samples and we reduce the number
of samples retained during the sample resolution period by 1.

3. Finally delphi remembers the index of the MAP sample withing the
retained samples.

Note: There are some (or more) repeated code in this push and previous
ones that needs to be cleaned up when there is more time.

* Fixed a place that gave a segmentation fault due to faulty logic

* Fixed an initialization bug

* Fixed a bug

The seasonality was not visible when Delphi was invoked through the
CauseMos interface. When a model is created, each concept’d period for
seasonality is inferred from data. However, it was not saved into the
serialized model.

When creating an experiment, an AnalysisGraph object is created by
deserializing the saved model. Since concept periods were not saved all
the concepts ended up having the default period of 1.

Co-authored-by: Manujinda Wathugala <manujinda@email.arizona.edu>