Delphi modeling 1 dev (#496)
* Accumulating log likelihoods

1. Remember the sequence of log likelihoods throughout the sampling
process including burning.

2. Complete state now includes the log likelihood sequence.

3. Plotting function now plots the log likelihoods.

* MAP Sample

1. Added code to identify and retain the MAP sample even if we encounter
during the burning period.

2. If we encounter the MAP sample within the burning period it is always
put at res - 1 position of the retained samples and we reduce the number
of samples retained during the sample resolution period by 1.

3. Finally delphi remembers the index of the MAP sample withing the
retained samples.

Note: There are some (or more) repeated code in this push and previous
ones that needs to be cleaned up when there is more time.

* Fixed a place that gave a segmentation fault due to faulty logic

* Fixed an initialization bug

* Fixed a bug

The seasonality was not visible when Delphi was invoked through the
CauseMos interface. When a model is created, each concept’d period for
seasonality is inferred from data. However, it was not saved into the
serialized model.

When creating an experiment, an AnalysisGraph object is created by
deserializing the saved model. Since concept periods were not saved all
the concepts ended up having the default period of 1.

* Generating synthetic CAGs

Added some methods to generate a random CAG and populate it with
synthetic observations.

* Cleanup

1. Made the code more modularized

2. Removed some debug print lines

* More cleaning

1. Removed commented code

2. Moved a block of code to make some repeated code identical in two
places. I did not refactor that into a separate method since it is only
a few lines of code and those few lines alone does not create a coherent
functionality.

* Created a class to perform timing

* Fixed CMakeLists.txt

* Changed a file name

* A basic macro profiler completed.

* Added some progress messages to the timer.

* Added proper error checking for opening delphi.db

This caused a chunk of repeated code. I feel that using a separate class
to handle the delphi.db is a better approach.

* Improved the newly created delphi.db open method to accept the database
open mode.

* Improved the Delphi macro timing program to accept command line options.

* Removed inclusion of dbg.h in main source files.

* Fixed the help option in timer.

* The june embed jupytre notebook got opened and some version number got chaned

* Seasonal curve from a Fourier series

The initial implementation of generating a cyclic curve using a LDS
based on Fourier analysis.

* Commented out some unused synthetic data generating methods.

Updated the timing program a bit.

* Deleted commented code related to synthetic data generation.

Co-authored-by: Manujinda Wathugala <manujinda@email.arizona.edu>