Added a stub to export a create model json file.
This should output a create model json file that would create the
untrained model of the object the method export_create_model_json_string()
is called upon.

This method would be mostly used to output synthetic CAGs with synthetic
data into create model json files so that we could run the training
pipeline on those models in a freshly created object other than the
object that was used to generate the synthetic CAG to prevent any
ground truth parameter values leaking from the synthetic CAG generation
process into the training process.

To aid assess the goodness of training, it would be beneficial to store
the ground truth parameters of the CAG that generated the data in a
separate additional section of the create model json produced.